pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-node-nccl-p2p-benchmarks"
version = "0.1.0"
description = "GPU P2P bandwidth matrix and CSV export for NVLink/PCIe diagnostics"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
p2p-bench = "p2p_bench.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
